requirements.txt
# Core
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# EEG / BCI
mne>=1.0.0
pyriemann>=0.3.0
moabb>=0.4.0

# Data / config
pandas>=1.3.0
h5py>=3.0.0
pyyaml>=5.4.0

# Viz
matplotlib>=3.4.0
seaborn>=0.11.0

# Utils
tqdm>=4.62.0
joblib>=1.1.0

# Tests
pytest>=7.0.0
pytest-cov>=3.0.0
